project hackdku/HacksonProject/requirements.txt
        Flask>=2.0
        Flask-Cors>=3.0 
        openai>=1.0  
        python-dotenv>=0.19
	pip install transformers torch torchvision torchaudio pillow accelerate